+++ CHANGELOG.md
@@ -7,6 +7,11 @@ WYMeditor.
 
 *release-date* TBD
 
+### Backwards-Incompatible Changes
+
+* [#613](https://github.com/wymeditor/wymeditor/pull/613)
+  ``editor.setSingleSelectionRange`` is no longer public API.
+
 ## 1.0.0-rc.1
 
 *release-date* October 24, 2014